Exceptions do not work if ClickHouse is integrated to AWS Lambda
The issue stems from a `mincore` call when checking the validity of the CFA pointer in `libunwind/src/DwarfInstructions.hpp:isPointerValid`. This system call is prohibited within the scope of the AWS Lambda environment, resulting in an `EPERM` error.

A workaround is to add a check for the `EPERM` errno to `isPointerValid` function. This disables CFA validity checks, but it appears to be the only viable option for AWS Lambda.